Modified project structure, finished readme
Modified project structure in order to build the redis and celery (celerybeat inclusive) as containers rather than needing to run workers through the commandline for ease of setup. This is reflected in the moved dockerfile and modified docker-compose. Requirements also moved as a result. Readme is now written with brief instructions on setup.